extract into form526_birls_ids.rb
