+++ app/models/concerns/form526_birls_ids.rb
@@ -1,0 +1,81 @@
+# frozen_string_literal: true
+
+module Form526BirlsIds
+  extend ActiveSupport::Concern
+
+  BIRLS_KEY = 'va_eauth_birlsfilenumber'
+
+  # this method is for queuing up BIRLS ids in the birls_ids_tried hash,
+  # and can also be used for initializing birls_ids_tried.
+  # birls_ids_tried has this shape:
+  # {
+  #   birls_id => [timestamp, timestamp, ...],
+  #   birls_id => [timestamp, timestamp, ...], # in practice, will be only 1 timestamp
+  #   ...
+  # }
+  # where each timestamp notes when a submissison job (start) was started
+  # with that BIRLS id (birls_id_tried keeps track of which BIRLS id
+  # have been tried so far).
+  # add_birls_ids does not overwrite birls_ids_tried.
+  # example:
+  # > sub.birls_ids_tried = { '111' => ['2021-01-01T0000Z'] }
+  # > sub.add_birls_ids ['111', '222', '333']
+  # > pp sub.birls_ids_tried
+  #    {
+  #      '111' => ['2021-01-01T0000Z'], # a tried BIRLS ID
+  #      '222' => [],  # an untried BIRLS ID
+  #      '333' => []   # an untried BIRLS ID
+  #    }
+  # NOTE: '111' was not cleared
+  def add_birls_ids(id_or_ids)
+    ids = Array.wrap(id_or_ids).map { |id| id.is_a?(Symbol) ? id.to_s : id }
+    hash = birls_ids_tried_hash
+    ids.each { |id| hash[id] ||= [] }
+    self.birls_ids_tried = hash.to_json
+    self.multiple_birls = true if birls_ids.length > 1
+    ids
+  end
+
+  def birls_ids
+    [*birls_ids_tried_hash&.keys, birls_id].compact.uniq
+  end
+
+  def birls_ids_tried_hash
+    birls_ids_tried.presence&.then { |json| JSON.parse json } || {}
+  end
+
+  def mark_birls_id_as_tried(id = birls_id!, timestamp_string: Time.zone.now.iso8601.to_s)
+    ids = add_birls_ids id
+    hash = birls_ids_tried_hash
+    hash[ids.first] << timestamp_string
+    self.birls_ids_tried = hash.to_json
+    timestamp_string
+  end
+
+  def mark_birls_id_as_tried!(*args, **kwargs)
+    timestamp_string = mark_birls_id_as_tried(*args, **kwargs)
+    save!
+    timestamp_string
+  end
+
+  def birls_ids_that_havent_been_tried_yet
+    add_birls_ids birls_id if birls_id.present?
+    birls_ids_tried_hash.select { |_id, timestamps| timestamps.blank? }.keys
+  end
+
+  def birls_id!
+    auth_headers[BIRLS_KEY]
+  end
+
+  def birls_id
+    birls_id! if auth_headers_json
+  end
+
+  def birls_id=(value)
+    headers = JSON.parse(auth_headers_json) || {}
+    headers[BIRLS_KEY] = value
+    self.auth_headers_json = headers.to_json
+    @auth_headers_hash = nil # reset cache
+  end
+
+end
