+++ app/models/form526_submission.rb
@@ -5,6 +5,7 @@ require 'sentry_logging'
 class Form526Submission < ApplicationRecord
   include SentryLogging
   include Form526Workflow
+  include Form526BirlsIds
 
   # A 526 disability compensation form record. This class is used to persist the post transformation form
   # and track submission workflow steps.
@@ -24,7 +25,7 @@ class Form526Submission < ApplicationRecord
   #     workflow complete.
   # @!attribute created_at
   #   @return [Timestamp] created at date.
-  # @!attribute workflow_complete
+  # @!attribute updated_at
   #   @return [Timestamp] updated at date.
   #
   has_kms_key
@@ -37,8 +38,6 @@ class Form526Submission < ApplicationRecord
   has_many :form526_job_statuses, dependent: :destroy
 
   validates(:auth_headers_json, presence: true)
-
-  BIRLS_KEY = 'va_eauth_birlsfilenumber'
 
   def start
     if single_issue_hypertension_claim? && Flipper.enabled?(:disability_hypertension_compensation_fast_track)
@@ -133,79 +132,6 @@ class Form526Submission < ApplicationRecord
   #
   def auth_headers
     @auth_headers_hash ||= JSON.parse(auth_headers_json)
-  end
-
-  # this method is for queuing up BIRLS ids in the birls_ids_tried hash,
-  # and can also be used for initializing birls_ids_tried.
-  # birls_ids_tried has this shape:
-  # {
-  #   birls_id => [timestamp, timestamp, ...],
-  #   birls_id => [timestamp, timestamp, ...], # in practice, will be only 1 timestamp
-  #   ...
-  # }
-  # where each timestamp notes when a submissison job (start) was started
-  # with that BIRLS id (birls_id_tried keeps track of which BIRLS id
-  # have been tried so far).
-  # add_birls_ids does not overwrite birls_ids_tried.
-  # example:
-  # > sub.birls_ids_tried = { '111' => ['2021-01-01T0000Z'] }
-  # > sub.add_birls_ids ['111', '222', '333']
-  # > pp sub.birls_ids_tried
-  #    {
-  #      '111' => ['2021-01-01T0000Z'], # a tried BIRLS ID
-  #      '222' => [],  # an untried BIRLS ID
-  #      '333' => []   # an untried BIRLS ID
-  #    }
-  # NOTE: '111' was not cleared
-  def add_birls_ids(id_or_ids)
-    ids = Array.wrap(id_or_ids).map { |id| id.is_a?(Symbol) ? id.to_s : id }
-    hash = birls_ids_tried_hash
-    ids.each { |id| hash[id] ||= [] }
-    self.birls_ids_tried = hash.to_json
-    self.multiple_birls = true if birls_ids.length > 1
-    ids
-  end
-
-  def birls_ids
-    [*birls_ids_tried_hash&.keys, birls_id].compact.uniq
-  end
-
-  def birls_ids_tried_hash
-    birls_ids_tried.presence&.then { |json| JSON.parse json } || {}
-  end
-
-  def mark_birls_id_as_tried(id = birls_id!, timestamp_string: Time.zone.now.iso8601.to_s)
-    ids = add_birls_ids id
-    hash = birls_ids_tried_hash
-    hash[ids.first] << timestamp_string
-    self.birls_ids_tried = hash.to_json
-    timestamp_string
-  end
-
-  def mark_birls_id_as_tried!(*args, **kwargs)
-    timestamp_string = mark_birls_id_as_tried(*args, **kwargs)
-    save!
-    timestamp_string
-  end
-
-  def birls_ids_that_havent_been_tried_yet
-    add_birls_ids birls_id if birls_id.present?
-    birls_ids_tried_hash.select { |_id, timestamps| timestamps.blank? }.keys
-  end
-
-  def birls_id!
-    auth_headers[BIRLS_KEY]
-  end
-
-  def birls_id
-    birls_id! if auth_headers_json
-  end
-
-  def birls_id=(value)
-    headers = JSON.parse(auth_headers_json) || {}
-    headers[BIRLS_KEY] = value
-    self.auth_headers_json = headers.to_json
-    @auth_headers_hash = nil # reset cache
   end
 
   class SubmitForm526JobStatusesError < StandardError; end
